refactor: split admin-api telemetry + content-engine triggers along DTO/handler and definition/matching seams
Pure code-movement split of two over-cap files (CLAUDE.md file-organization
hard cap of 700 lines), per tech-debt issue #529.

admin-api telemetry.rs (836) -> telemetry/:
  - mod.rs      module doc, size-cap consts, routes()
  - dto.rs      wire types (TelemetryEvent + variants, Chunk/BundleResponse,
                IngestError + IntoResponse, AuthError::clone_status_response)
  - handlers.rs upload_chunk/upload_bundle + unpack_and_replay/verify_bearer/
                replay_event/replay_client_native
  - tests.rs    moved #[cfg(test)] module

content-engine triggers.rs (800) -> triggers/:
  - mod.rs      Trigger / TriggerType / TriggerEvent definitions
  - matching.rs impl Trigger { trigger_type(), matches() }
  - tests.rs    moved #[cfg(test)] module

Behavior-preserving - re-export discipline - zero public-surface change.
DTO struct fields widened to pub *within* pub(super) types so the sibling
handlers/tests modules can read them; the types stay pub(super), so external
visibility is unchanged. dev_session helpers reached via crate::routes::
dev_session (absolute) instead of super:: to absorb the extra module depth.

Part of #529
